vcli-core: integration smoke test — fixture canonical + hash stable
Add tests/fixture_canonical.rs with two smoke tests verifying that the
YT ad skipper fixture canonicalizes and hashes stably across parse +
re-serialize.

Fixes required to make round-trip canonical equality hold:
- Confidence: f32 → f64 (preserves JSON precision; 0.9 no longer becomes
  0.8999999761581421 after struct round-trip)
- Step::Click button: add skip_serializing_if for the Left default so the
  re-serialized form omits the implicit field, matching the fixture
- Program::body: remove skip_serializing_if so empty body always serializes,
  keeping key set stable across raw-JSON → Program → serde_json::Value
- canonical.rs sha256_hex: allow(too_many_lines) for inline SHA-256 impl

+++ crates/vcli-core/src/predicate.rs
@@ -10,11 +10,12 @@ use crate::clock::UnixMs;
 use crate::geom::Rect;
 use crate::region::Region;
 
-/// Confidence as a `f32` in [0, 1]. Newtyped to avoid silent mismatches with
-/// arbitrary f32s and to prevent accidental equality checks that surprise.
+/// Confidence as a `f64` in [0, 1]. Newtyped to avoid silent mismatches with
+/// arbitrary floats and to prevent accidental equality checks that surprise.
+/// Uses `f64` (not `f32`) to preserve the full JSON precision on round-trip.
 #[derive(Debug, Clone, Copy, PartialEq, PartialOrd, Serialize, Deserialize)]
 #[serde(transparent)]
-pub struct Confidence(pub f32);
+pub struct Confidence(pub f64);
 
 /// RGB triple for color matching.
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
@@ -59,7 +60,7 @@ pub enum PredicateKind {
         /// hash at eval time and compares).
         baseline: String,
         /// Fractional Hamming-distance threshold (0.0..=1.0).
-        threshold: f32,
+        threshold: f64,
     },
     /// Logical AND across named predicates.
     AllOf {
